Add script and make rule to setup output folders
We want the root `output/` folder to be the canonical location, with
suitable subfolders (`output/AC/`, etc.) to hold each individual book.
That's because generating the whole set of books is the default.

However, we also want to preserve the ability to generate individual
books.  Each individual book has its own Makefile, and should work
correctly independently of the others.  It would be inappropriate for a
subfolder Makefile to reach into a higher-level directory to make the
canonical output folder and symlink to it.  Thus, whenever we make
output rules for those Makefiles, they will be simple `mkdir` commands.

In anticipation of this, and to make it so that Makefiles can be run in
any order, we add a script that migrates any content in a subfolder
output directory to the canonical location, and replaces it with a
symlink.  We force this rule to run every time someone requests the
`output_folders` target, because it's idempotent and cheap.

We also add the output folders to our `.gitignore`, since we should be
able to blow away their contents at will and regenerate them.

Helps #8.