Remove tar/zip creating rules
Rationale: the maintenance burden of these rules is no longer justified,
because people who want to contribute should just clone the git repo.
There was some value in zipping up the generated versions of converted
images, but the switch to `pdflatex` lets us just use `.png` and `.jpg`
files directly, removing the most error prone part of the process.

As a side benefit, we're able to remove the clumsy, oft-misused,
everything-and-the-kitchen-sink `SRC` variables, in favor of explictly
declared dependencies.

Fixes #9.