Change repeated reference stem
I see no value in permitting different references with the same stem.
The cost, though, is that it's incompatible with the new policy where we
remove the extension and keep only the stem: it causes multiply-defined
references.  Fortunately, there is only one instance of this in the
codebase presently, so removing it is a simple solution.

Helps #12.